Update and rename database_test.go to preimages_test.go
